pyproject.toml
[project]
name = "bro-code"
version = "0.2.0"
description = "AI-powered multi-agent coding assistant with web UI"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.40.0",
    "langchain>=0.3.27",
    "langchain-core>=0.3.72",
    "langchain-anthropic>=0.3.0",
    "langgraph>=0.6.3",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
    "fastapi>=0.115.0",
    "uvicorn>=0.34.0",
    "sse-starlette>=2.2.1",
]
